backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
python-dotenv>=1.0.0
openai>=1.12.0
httpx>=0.26.0
apscheduler>=3.10.4
pydantic>=2.8.0
pydantic-settings>=2.1.0

# Note: If you encounter build errors with pydantic-core on Python 3.14,
# install packages without version constraints:
# pip install fastapi uvicorn[standard] sqlalchemy python-dotenv openai httpx apscheduler pydantic pydantic-settings

